+++ NEWS.md
@@ -1,6 +1,8 @@
-* Added additional tests for bug fixes.
-    * Now up to 769 tests, all passing.
+* Added additional tests for bug fixes and new features.
+    * Now up to 782 tests, all passing.
     * Code coverage remains at 100%.
+* Added new function `rename_to_pref_suff_byname()` which 
+  renames rows or columns to prefixes or suffixes in row and column names.
 * Fixed a bug in `clean_byname()` which caused 
   a `NULL` response when unnamed rows or columns were present.
 * Now using new `prepare_.FUNdots()` function in all `*apply_byname()` functions.